requirements.txt
# Player Churn Prediction System

# Core dependencies
pandas==2.1.4
numpy==1.26.2
scikit-learn==1.3.2
xgboost==2.0.3
lightgbm==4.1.0
shap==0.44.0

# Data processing
scipy==1.11.4
pyarrow==14.0.1

# Database
psycopg2-binary==2.9.9
redis==5.0.1

# API
fastapi==0.108.0
uvicorn[standard]==0.25.0
pydantic==2.5.3
python-multipart==0.0.6

# Airflow
apache-airflow==2.8.0
apache-airflow-providers-postgres==5.10.0

# Visualization
streamlit==1.29.0
plotly==5.18.0
matplotlib==3.8.2
seaborn==0.13.0

# Monitoring
grafana-client==3.5.0

# Utilities
python-dotenv==1.0.0
requests==2.31.0
tqdm==4.66.1

# Testing
pytest==7.4.3
pytest-cov==4.1.0
pytest-mock==3.12.0

# Development
jupyter==1.0.0
ipykernel==6.27.1
black==23.12.1
flake8==7.0.0
mypy==1.7.1

# Kafka (optional, for real-time streaming)
kafka-python==2.0.2

# Documentation
mkdocs==1.5.3
mkdocs-material==9.5.3
